Updated code for merge
