--- keyboards/macroflow_original/keymaps/default/config.h
@@ -1,6 +1,0 @@
-// Copyright 2024 Patrick Mathern (@Patrickemm)
-// SPDX-License-Identifier: GPL-2.0-or-later
-
-#define DYNAMIC_KEYMAP_LAYER_COUNT 8
-#define ENCODER_RESOLUTION 4
-#define OLED_TIMEOUT 60000
