+++ keyboards/macroflow_original/keymaps/default/rules.mk
@@ -1,3 +1,1 @@
-VIA_ENABLE = yes
 ENCODER_MAP_ENABLE = yes
-OLED_ENABLE = yes
